Factor out GPUMemAllocator into its own include.
This allows using GPUMemAllocator without pulling in TF protobufs as a dependency.

PiperOrigin-RevId: 248040160
